+++ apps/expo/app/index copy.tsx
@@ -1,0 +1,5 @@
+import { HomeScreen } from 'app/features/home/screen'
+
+export default function Home() {
+    return <HomeScreen />
+}
