+++ apps/expo/app/index.tsx
@@ -1,5 +1,5 @@
-import { HomeScreen } from 'app/features/home/screen'
+import { DesignShowcase } from '@my/ui/src/components/DesignShowcase'
 
 export default function Home() {
-    return <HomeScreen />
+  return <DesignShowcase />
 }
